refactor(shared): update API client imports to use shared/api
