+++ src/effect/gatherings/leaveGathering.ts
@@ -1,4 +1,4 @@
-import client from "../client/client";
+import { client } from "@/shared/api";
 
 // 모임 참여 취소하기
 export const leaveGathering = async (id: number) => {
