+++ src/effect/reviews/createReview.ts
@@ -1,4 +1,4 @@
-import client from "../client/client";
+import { client } from "@/shared/api";
 
 interface CreateReviewRequest {
   gatheringId: number;
